Rename and make tests compile.
