+++ src/client/mod.rs
@@ -1,2 +1,1 @@
 pub mod editor_client;
-pub use editor_client::*;
