+++ src/lib.rs
@@ -3,8 +3,3 @@ pub mod text_buffer;
 pub mod syntax;
 pub mod server;
 pub mod client;
-
-pub use server::*;
-pub use client::*;
-pub use text_buffer::*;
-pub use syntax::*;
