+++ src/main.rs
@@ -6,24 +6,30 @@ mod editor;
 mod text_buffer;
 mod syntax;
 mod server;
-
-fn main() -> io::Result<()> {
+mod client;
+#[tokio::main]
+async fn main() -> io::Result<()> {
     let args: Vec<String> = std::env::args().collect();
 
-    // let mut editor = if args.len() > 1 {
-    //     let path = PathBuf::from(&args[1]);
-    //     match Editor::open_file(path).await {
-    //         Ok(editor) => editor,
-    //         Err(e) => {
-    //             eprintln!("Error opening file: {} {}", &args[1], e);
-    //             std::process::exit(1);
-    //         }
-    //     }
-    // } else {
-    //     Editor::new()
-    // };
-    //
-    // match editor.run() {
+    let mut editor = if args.len() > 1 {
+        let path = PathBuf::from(&args[1]);
+        match Editor::open_file(path).await {
+            Ok(editor) => editor,
+            Err(e) => {
+                eprintln!("Error opening file: {} {:?}", &args[1], e);
+                std::process::exit(1);
+            }
+        }
+    } else {
+        match Editor::new().await {
+            Ok(editor) => editor,
+            Err(e) => {
+                eprintln!("Error creating editor file: {} {:?}", &args[1], e);
+                std::process::exit(1);
+            }
+        }    };
+
+    // match editor.run().await {
     //     Ok(()) => {
     //         println!("editor exited successfully");
     //     }
@@ -31,6 +37,6 @@ fn main() -> io::Result<()> {
     //         println!("editor exited with error: {}", e);
     //     }
     // }
-    //
+
     Ok(())
 }
